feat: implemented diff_similarity using the hora crate
hora supports a number of approximate nearest neighbor methods. From a usability perspective, it is not good for what we want to achieve, because it requires a normalization of git diffs to float vectors. Furthermore, the search does not allow targeting certain similarities. Too many results are returned, and there is no easy way to filter them. In conclusion, we should look for alternatives.