feat: implemented first shingling function for commit diffs
Shingling is performed line-wise. This means that each shingle consists of _n_ lines, where _n_ is the arity of the shingling process. All lines are trimmed by removing leading and trailing whitespace.

The lines in a shingle are simply concatenated without line separators. I made this design choice, because line separators do not carry useful information for the hashing process that follows.