feat: implemented first one-hot encoding for shingled diffs
The one-hot encoding is done with the help of a vocabulary that tracks all encountered words (i.e. shingles). The vocabulary assigns a random index to each word. A given text or diff is then encoded by creating a vector of the size of the vocabulary, where a value at a certain index is set to '1', if the text/diff contains the word with this index, otherwise the value is '0'.